gw get visit by id
